python/requirements.txt
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
disruptive==1.5.5
executing==1.2.0
fonttools==4.38.0
grpcio==1.51.1
grpcio-tools==1.51.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
lnetatmo==2.1.0
matplotlib==3.7.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
protobuf==4.22.0
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pywin32==305
pyzmq==25.0.0
requests==2.28.2
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
zipp==3.14.0
